requirements.txt
Django==4.1.4
django-auditlog==2.2.1
django-filter==22.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
psycopg2-binary==2.9.5
coreapi==2.3.3
pillow==9.3.0
django-environ==0.9.0
factory_boy==3.2.1
faker==15.3.4
gunicorn==20.1.0
django-cors-headers==3.13.0
sentry-sdk==1.12.1
drf-recaptcha==2.2.2